web-scraping/spotify/requirements.txt
python-dotenv==0.20.0
spotipy==2.19.0
